Create Env_check_outlook_email
This script will check all the environment URL's to make sure all QA or any related environments are up and running. After execution it will email the results in a html table via outlook email with running environments highlighted in Green colour and failed ones in Red colour (for better readability).
I also scheduled this script to run at specific time everyday and email automatically to all the recipients added in script (scheduling part is not covered in this script).